dark mode

+++ android_client/app/src/main/java/com/example/android_client/repositories/DataRepository.java
@@ -80,6 +80,9 @@ public class DataRepository {
     }
 
     public void init(LifecycleOwner lifecycleOwner, MutableLiveData initialized) {
+        if(initialized.isInitialized()){
+            return;
+        }
         Thread deleteThread = new Thread(() -> {
             try {
                 userDao.deleteAll();
